requirements.txt
python-dotenv==1.1.0
python-datauri==3.0.2
loguru==0.7.3
minio==7.2.15
kafka-python==2.1.4
pydantic==2.11.1
asgiref==3.8.1
langchain-text-splitters==0.3.8
langchain==0.3.22
langchain-pinecone==0.2.3
langchain-openai==0.3.12
aiokafka==0.12.0
fastapi==0.115.12
uvicorn==0.34.0
google-cloud-documentai==3.3.0
PyPDF2==3.0.1
